fix(LottoRepository) : 나눗셈 과정 중 Double형 맞추지 않는 오류 수정

+++ src/main/kotlin/lotto/domain/repository/LottoRepository.kt
@@ -15,7 +15,7 @@ class LottoRepository {
     }
 
     fun calEarningRate(purchasePrize: Money, totalPrice: Long): Double {
-        return (totalPrice / purchasePrize.getMoney() * 100.0)
+        return (totalPrice / purchasePrize.getMoney().toDouble() * 100.0)
     }
 
     fun calStatistics(winningLottery: WinningLottery, purchaseLottos: List<Lotto>): List<Rank> {
